move details carousel item into its own file
